internal_solver_solv: do not return if provider is already installed
If package 'a' Provides 'd' and is installed, and 'b' also Provides 'd'
and is available, when:

	opkg install 'd'

'd' should be able to be installed. The unnecessary
is_provides_installed check is currently incorrectly succeeding since it
only checks for providers that match the package name, instead of
also checking the Provides field of the package.

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>