libsolv_solver_execute_transaction: add case for obsolete packages
Given a system that have package 'a' installed, where both package 'a'
and 'b' Replaces/Conflicts/Provides 'v'. Installing package 'b' should
remove 'a' first. Currently, that operation is expressed as install b,
remove a, which end up running the maintainer scripts in the wrong
order. Add code to treat the Obsolete operation separetely, as a remove
a, install b.

bugzilla #11362

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>