libsolv_solver_transaction_preamble: add arch information to get matching package
The function pkg_hash_fetch_by_name_version is called to find the pkg
object that corresponds to the solvable returned by libsolv. The
matching is done using name and version, which is incomplete as there
might be several packages with the same name/version, but with different
architecture. Add architecture information to the comparision.

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>