remove_maintainer_scripts: use strict matching
The function is using a glob to select which metadata files needs to be
deleted during package removal, on the info_dir. However, the glob may
match metadata files from packages with similar names. For example,
during removal of package glibc-binary-localedata-de-at, the current
logic was also removing the metadata for
glibc-binary-localedata-de-at.iso-8859-1. Add check for an exact match
before deletion.

Fixes bugzilla: 12905

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>