Initialize pkg->tags
From bugzilla 13380, authored by lnksz.dev@gmail.com [1]:

Freeing an uninitialized pointer.

```c
pkg_t pkg;
pkg_init_from_file(&next, "/tmp/some.ipk"); // This calls pkg_init() under the hood
// Use pkg here
pkg_deinit(&pkg); // ==> free(pkg->tags) freeing an uninitialized pointer
```

This is not surfacing if one uses the pkg_new() call, as the underlaying
calloc() sets this to 0, which could be NULL, so the free is ok. But as
the described method is also legit by the API and spares a possible
memory leak, the init should be added.

[1] https://bugzilla.yoctoproject.org/show_bug.cgi?id=13380

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>