opkg: add verify option
Add verify option to verify the current md5sum of installed files vs the
md5sum of the file at install time. The check is mainly aimed as an
integrity check, to detect which files has been modified. It is not
meant to be a secure way to determine if a system has been tampered
with.

The verify option relies on checksums stored on a metadata file
(md5sums)  which needs to be packaged with the rest of the control files
on the control tarball. The file has key/value pairs with
md5sum/filepath data. The file can be generated at build time (before
running opkg-build, if that's the tool being used). An example command
that can generate the md5sums file:

find . -type f -regex '.*?debian-binary.*' ! -regex '.*?CONTROL.*'
-printf '%P ' | xargs md5sum > CONTROL/md5sums

This file is optional, if present it will be used with the --verify
option, if not, the operation will be a no-op.

This command is largely model after the dpkg -V option.

Bugzilla #7649

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>