libsolv_solver_execute_transaction: treat obsolete operations as upgrades
When the Replace/Conflict pattern is used to split a package [1], opkg is
configuring libsolv to return replace operations in two steps. This
works well for a package that is split in 2, but falls apart when a
package is split into more than 2 packages. Insted, instruct libsolv to
treat OBSOLETE operations as upgrades.

[1] https://www.debian.org/doc/debian-policy/ch-relationships.html

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>