tests/opk.py: enable subdirectories in test feeds
Add an optional `subdirectory` kwarg to the Opk class which allows tests
to create opkg packages under subdirectories within the test feed. This
is useful when sketching out subfeed architectures.

The subdirectories are created if they are not already present, and
their relative paths to the Packages index file are included in the
package's `Filename` attribute.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>
Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>