libopkg: do not require uid 0 for RO cmds
The current opkg_conf implementation acquires an exclusive lock in the
`/var/run/` directory, when it is initialized. As a result, all opkg
operations must be run as a user which has write access to that lock
directory - which should always be uid 0.

Instead, move the lock context to the opkg_cmd execution function and
add an attribute to the opkg_cmd_t which expresses whether or not the
command requires privileged execution. For all operations which do not
modify the system state, lock capture is eschewed.

When a non-root user tries to execute a privileged command, they will
receive an EPERM.

Fixes bugzilla #9348

Signed-off-by: Alex Stewart <alex.stewart@ni.com>
Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>