opkg_conf: reset lock_fd after unlinking file
The lock file descriptor is not reset to its NULL-signalling value (-1)
after it is unlinked. As a result, calling _unlock() multiple times in
succession will fail in unintuitive ways, as the function attempts to
unlink a dead descriptor.

Reset the value of the lock_fd after the lock file has been successfully
unlinked, as a signal to later _unlock() calls to NOOP.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>
Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>