opkg_cmd: make download return code more useful
The return values of the opkg_download_cmd and opkg_install_cmd differ
appreciably. When any requested package in an install operation is not
solvable (eg. it is not present in the index), install returns -1
(return code 255 in the shell). If the download operation encounters the
same situation, it ignores the package and returns code 0.

As a result, there is no reliable way to determine if a download
operation was unsuccessful due to a missing package.

Set the return code on the opkg_download_cmd operation to -1 in the case
that any package cannot be found in the index.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>
Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>