tests: let the OS negotiate relative package dirs
In cases where a regression test requires that a package be installed to
a subdirectory in the test feed, the opk.py module will attempt to
resolve the subdirectory and rebase it to the root of the test feed.

This is unnecessary, since all operations which make use of the
subdirectory path do so from the perspective of the test feed already.
Further, the rebase implementation breaks in cases where the test feed
is beyond a symlink.

Remove the resolve-and-rebase logic, and allow the OS to negotiate the
relative path.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>